Moved TexifyUtil#getReferencedFiles to FilesKt.
